WIP visualizer class to test with real codebase
This is the start of the rewrite of the visualizer. A lot of traversing logic has been copied over from the old one, and visualizing logic has been deleted for now. I'm going through a real-world codebase with this to check and resolve parsing failures to discover most possible use cases.